requirements.txt
django-composite-foreignkey==1.1.0
django-bootstrap-v5==1.0.11
django-bootstrap4==23.2
django-composite-foreignkey==1.1.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-formtools==2.4.1
django-viewflow==2.0.0b6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
